requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
jsonschema==4.20.0
cryptography==41.0.7
requests==2.31.0
python-multipart==0.0.6
jsonpath-ng==1.6.1
python-dotenv==1.0.0
